Moved interfaces to their implementations
